bank-etl-migration/requirements.txt
pandas
psycopg2-binary
python-dateutil
pyyaml
tqdm 
